refactor: index.js перемещен в pages, constants.js перемещен в папку utils
